feat: replicator suffix test
