+++ packages/cohort/src/replicator/tests/mod.rs
@@ -1,0 +1,3 @@
+pub mod suffix;
+pub mod test_utils;
+pub mod utils;
